% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_desc.R
\name{agree_index}
\alias{agree_index}
\title{Calculate Agreement statistic}
\usage{
agree_index(data, items, cut_score, scale)
}
\arguments{
\item{data}{A data frame of dichotomously scored test times}

\item{items}{Raw column indices representing the test items}

\item{cut_score}{A raw or percentage cut-score}

\item{scale}{A character vector indicataing wheter the cut-score
is 'raw' (default) or 'percent'}
}
\value{
Agree Agreement statistic values for items on the test
}
\description{
Calculate Agreement statistic
}

