% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_desc.R
\name{agree_stat}
\alias{agree_stat}
\title{Calculate Agreement statistic}
\usage{
agree_stat(data, items, cut_score, scale = "raw")
}
\arguments{
\item{data}{A data frame of dichotomously scored test times}

\item{items}{Raw column indices representing the test items}

\item{cut_score}{A raw or percentage cut-score}

\item{scale}{A character vector indicating whether the cut-score
is 'raw' (default) or 'percent'}
}
\value{
Agree Agreement statistic values for items on the test
}
\description{
Calculate Agreement statistic
}
\examples{
agree_stat(bh_depend, 2:31, 21, scale = 'raw')
}
