% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_desc.R
\name{b_index}
\alias{b_index}
\title{Calculate B-index}
\usage{
b_index(data, items, cut_score, scale = "raw")
}
\arguments{
\item{data}{A data frame of dichotomously scored test times}

\item{items}{Raw column indices representing the test items}

\item{cut_score}{A raw or percentage cut-score}

\item{scale}{A character vector indicating whether the cut-score
is 'raw' (default) or 'percent'}
}
\value{
Bindex B-index values for items on the test
}
\description{
Calculate B-index
}
\examples{
b_index(bh_depend, 2:31, 21, scale = 'raw')
}
