% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bh_depend}
\alias{bh_depend}
\title{Brown and Hudson's (2002, p. 153) Table 5.1a Data for 45 examinees on 30 test items}
\format{A data frame with 45 rows and 30 variables

\itemize{
\item ID. ID numbers.
\item L_1. Listening question 1.
\item L_2. Listening question 2.
\item L_3. Listening question 3.
\item L_4. Listening question 4.
\item L_5. Listening question 5.
\item L_6. Listening question 6.
\item L_7. Listening question 7.
\item L_8. Listening question 8.
\item L_9. Listening question 9.
\item L_10. Listening question 10.
\item R_1. Reading question 1.
\item R_2. Reading question 2.
\item R_3. Reading question 3.
\item R_4. Reading question 4.
\item R_5. Reading question 5.
\item R_6. Reading question 6.
\item R_7. Reading question 7.
\item R_8. Reading question 8.
\item R_9. Reading question 9.
\item R_10. Reading question 10.
\item G_1. Grammar question 1.
\item G_2. Grammar question 2.
\item G_3. Grammar question 3.
\item G_4. Grammar question 4.
\item G_5. Grammar question 5.
\item G_6. Grammar question 6.
\item G_7. Grammar question 7.
\item G_8. Grammar question 8.
\item G_9. Grammar question 9.
\item G_10. Grammar question 10.
}}
\source{
Brown and Hudson (2002)
}
\usage{
data(bh_depend)
}
\description{
A data set containing the 45 scored responses to 30 items on a
test of Listening (1-10), Reading (11-20), and Grammar (20-30)
}
\keyword{datasets}
