% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bh_item}
\alias{bh_item}
\title{Brown and Hudson's (2002, p. 124) Table 4.8 Item analysis data (first 10 items only)}
\format{A data frame with 15 rows and 12 variables

\itemize{
\item Students. Student names.
\item Q1. Question 1.
\item Q2. Question 2.
\item Q3. Question 3.
\item Q4. Question 4.
\item Q5. Question 5.
\item Q6. Question 6.
\item Q7. Question 7.
\item Q8. Question 8.
\item Q9. Question 9.
\item Q10. Question 10.
\item Total. Total score.
}}
\source{
Brown and Hudson (2002)
}
\usage{
data(bh_item)
}
\description{
A data set containing the scored responses to first 10 items of a
test and the total scores for 15 people
}
\keyword{datasets}
