% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_desc.R
\name{crt_iteman}
\alias{crt_iteman}
\title{Calculate item discrimination indices}
\usage{
crt_iteman(data, items, cut_score, scale = "raw")
}
\arguments{
\item{data}{A data frame of dichotomously scored test times}

\item{items}{Raw column indices representing the test items}

\item{cut_score}{A raw or percentage cut-score}

\item{scale}{A character vector indicataing wheter the cut-score
is 'raw' (default) or 'percent'}
}
\value{
B_index B-index values for items on the test

Agree_stat Agreement statistic values for items on the test

Item_Phi Item Phi values for items on the test
}
\description{
Calculate item discrimination indices
}

