% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_desc.R
\name{IF}
\alias{IF}
\title{Calculate item facility}
\usage{
IF(data, items)
}
\arguments{
\item{data}{A data frame of dichotomously scored test times}

\item{items}{Raw column indices representing the test items}
}
\value{
Item_facility Item facility values for test items
}
\description{
Calculate item facility
}

