% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_desc.R
\name{item_phi}
\alias{item_phi}
\title{Calculate Item Phi}
\usage{
item_phi(data, items, cut_score, scale = "raw")
}
\arguments{
\item{data}{A data frame of dichotomously scored test times}

\item{items}{Raw column indices representing the test items}

\item{cut_score}{A raw or percentage cut-score}

\item{scale}{A character vector indicating whether the cut-score
is 'raw' (default) or 'percent'}
}
\value{
Phi Item Phi values for items on the test
}
\description{
Calculate Item Phi
}
\examples{
item_phi(bh_depend, 2:31, 21, scale = 'raw')
}
