% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependability.R
\name{short_phi}
\alias{short_phi}
\title{Calculate Brown's short-cut estimate for domain score dependability}
\usage{
short_phi(data, items)
}
\arguments{
\item{data}{A data frame of dichotomously scored test items}

\item{items}{Raw column indices representing the test items}
}
\value{
The \code{phi} estimate for domain score dependability
}
\description{
Calculate Brown's short-cut estimate for domain score dependability
}
\examples{
short_phi(brown_depend, 2:31)
}

