% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sub_agree_coef}
\alias{sub_agree_coef}
\title{Subkoviak's (1988) table of approximate values of the agreement coefficient}
\format{A data frame with 21 rows and 10 columns

\itemize{
\item z. z cut-score for test.
\item r_0.1. Reliability of 0.1.
\item r_0.1. Reliability of 0.2.
\item r_0.1. Reliability of 0.3.
\item r_0.1. Reliability of 0.4.
\item r_0.1. Reliability of 0.5.
\item r_0.1. Reliability of 0.6.
\item r_0.1. Reliability of 0.7.
\item r_0.1. Reliability of 0.8.
\item r_0.1. Reliability of 0.9.
}}
\source{
Subkoviak (1988)
}
\usage{
data(sub_agree_coef)
}
\description{
Subkoviak's (1988) table of approximate values of the agreement coefficient
}
\keyword{datasets}
