% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_desc.R
\name{if_pass}
\alias{if_pass}
\title{Calculate item facility for passing students}
\usage{
if_pass(data, items, cut_score, scale = "raw")
}
\arguments{
\item{data}{A data frame of dichotomously scored test times}

\item{items}{Raw column indices representing the test items}

\item{cut_score}{A raw or percentage cut-score}

\item{scale}{A character vector indicating whether the cut-score
is 'raw' (default) or 'percent'}
}
\value{
Item_facility_pass Item facility values for test items of
of test takers who passed the test
}
\description{
Calculate item facility for passing students
}
\examples{
if_pass(bh_depend, 2:31, 21, scale = 'raw')
}
