% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_desc.R
\name{if_total}
\alias{if_total}
\title{Calculate item facility}
\usage{
if_total(data, items)
}
\arguments{
\item{data}{A data frame of dichotomously scored test times}

\item{items}{Raw column indices representing the test items}
}
\value{
Item_facility Item facility values for test items
}
\description{
Calculate item facility
}
\examples{
if_total(bh_depend, 2:31)
}
