% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependability.R
\name{phi_domain}
\alias{phi_domain}
\title{Calculate Brennan's short-cut estimate for domain score dependability}
\usage{
phi_domain(data, items, n_items = NULL)
}
\arguments{
\item{data}{A data frame of dichotomously scored test items}

\item{items}{Raw column indices representing the test items}

\item{n_items}{Number of items on the test (needed if data does not contain item-level information)}
}
\value{
The \code{phi} estimate for domain score dependability
}
\description{
Calculate Brennan's short-cut estimate for domain score dependability
}
\examples{
phi_domain(bh_depend, 2:31)
}
