% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependability.R
\name{phi_domain}
\alias{phi_domain}
\title{Calculate Brown's (1990) short-cut estimate for phi dependability}
\usage{
phi_domain(data, items, total = NULL)
}
\arguments{
\item{data}{A data frame of dichotomously scored test items}

\item{items}{Raw column indices representing the test items or 
number of items on the test (see Details).}

\item{total}{Total score column name of the test (see Details)}
}
\value{
The \code{phi} estimate for domain score dependability.
}
\description{
Calculate Brown's (1990) short-cut estimate for phi dependability
}
\details{
When the item-level information is available, Kuder-Richardson 20 is used
as an estimate of alpha. If only the total scores on the test are available and the
number of items is known, Kuder-Richardson 21 is used as an estimate of alpha.
}
\examples{
phi_domain(bh_depend, 2:31)
}
