% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependability.R
\name{phi_lambda}
\alias{phi_lambda}
\title{Calculate Brennan's short-cut estimate for phi lambda}
\usage{
phi_lambda(data, items, scores, cut_score)
}
\arguments{
\item{data}{A data frame of dichotomously scored test items}

\item{items}{Number of items on the test (needed if data does not contain item-level information)}

\item{scores}{Column name of raw test scores}

\item{cut_score}{Cut-score of the test expressed as a proportion (e.g., 0.70)}
}
\value{
The phi lambda estimate for dependability
}
\description{
Calculate Brennan's short-cut estimate for phi lambda
}
\examples{
phi_lambda(data = bh_item, items = 100, scores = "Total", cut_score = 0.70)
}
