% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependability.R
\name{phi_lambda}
\alias{phi_lambda}
\title{Calculate Brennan's (1984) estimate for phi lambda}
\usage{
phi_lambda(data, items, cut_score, total = NULL)
}
\arguments{
\item{data}{A data frame of dichotomously scored test items}

\item{items}{Raw column indices representing the test items or 
number of items on the test}

\item{cut_score}{Cut-score of the test expressed as a proportion (e.g., 0.70)}

\item{total}{Column name of raw test scores.}
}
\description{
Calculate Brennan's (1984) estimate for phi lambda
}
\examples{
phi_lambda(data = bh_item, items = 100, total = "Total", cut_score = 0.70)
}
