% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependability.R
\name{subkoviak}
\alias{subkoviak}
\title{Calculate Subkoviak's (1988) single administration consistency indices}
\usage{
subkoviak(data, items, raw_cut_score, look_up = FALSE)
}
\arguments{
\item{data}{A data frame of dichotomously scored test items}

\item{items}{Raw column indices representing the test items}

\item{raw_cut_score}{The raw cut-score for the test}

\item{look_up}{If TRUE, the agreement and kappa tables from Subkoviak (1988) are returned with the results}
}
\value{
The \code{z_cut} score and the rounded \code{z_cut_rounded} score for the test and rounded values for the \code{agree_coef} (agreement) and \code{kappa_coef} (kappa) coefficients from Subkoviak's (1988) tables

The \code{z_cut} score and the rounded \code{z_cut_rounded} score for the test and rounded values for the \code{agree_coef} (agreement) and \code{kappa_coef} (kappa) coefficients from Subkoviak's (1988) tables

The \code{z_cut} score and the rounded \code{z_cut_rounded} score for the test and rounded values for the \code{agree_coef} (agreement) and \code{kappa_coef} (kappa) coefficients from Subkoviak's (1988) tables
}
\description{
Calculate Subkoviak's (1988) single administration consistency indices
}
\examples{
subkoviak(brown_depend, 2:31, 21)
}

